Drop python-qpid
A bot reported last week that Manila got on the orphan list in
F23 Rawhide. This is because python-qpid is orphaned (built from
qpid-cpp, which is actually orphaned). While waiting for this to
be fixed, why do we even depend on this stuff? We pull the
python-oslo-messaging and that ought to be enough.

Vladan Popovic commited this change to Nova on April 11, 2014 -
more than a year ago (288dd81569ba16a5aac099e04abdf2e5a443971d).