Update patch set 1
Patch Set 1: Code-Review-1

I'm not against this, but then we need to make sure that the deploymento tools take care of installation of lvm2 and/or samba when needed. Taking a look into puppet-manila i don't find installation of them so i guess that should be fixed in advance.

Patch-set: 1
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1